+++ astroquery/linelists/jplspec/core.py
@@ -4,11 +4,10 @@ import warnings
 
 import astropy.units as u
 from astropy.io import ascii
-from ..query import BaseQuery
-from ..utils import async_to_sync
+from astroquery.query import BaseQuery
+from astroquery.utils import async_to_sync
 # import configurable items declared in __init__.py
-from . import conf
-from . import lookup_table
+from astroquery.linelists.jplspec import conf, lookup_table
 from astroquery.exceptions import EmptyResponseError, InvalidQueryError